chore(Validation): added parameter validation on controllers and better error messages.

+++ src/main/java/br/com/lucashilles/domains/patient/PatientController.java
@@ -2,9 +2,6 @@ package br.com.lucashilles.domains.patient;
 
 import br.com.lucashilles.domains.reading.Reading;
 import br.com.lucashilles.domains.reading.ReadingProjection;
-import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
-import org.eclipse.microprofile.openapi.annotations.media.Schema;
-import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
 import org.jboss.resteasy.annotations.jaxrs.PathParam;
 import org.jboss.resteasy.annotations.jaxrs.QueryParam;
 
@@ -15,6 +12,9 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static br.com.lucashilles.domains.message.GenericMessage.ENTITY_NOT_FOUND;
+import static br.com.lucashilles.domains.message.GenericMessage.ID_MUST_BE_POSITIVE;
 
 @Path("/patient")
 @Produces(MediaType.APPLICATION_JSON)
@@ -41,6 +41,10 @@ public class PatientController {
     @Path("/{id}")
     @Transactional
     public Response deleteById(@PathParam long id) {
+        if (id < 0) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(ID_MUST_BE_POSITIVE).build();
+        }
+
         patientService.delete(id);
         return Response.ok("Patient deleted.").build();
     }
@@ -48,22 +52,38 @@ public class PatientController {
     @GET
     @Path("/{id}")
     public Response getById(@PathParam long id) {
+        if (id < 0) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(ID_MUST_BE_POSITIVE).build();
+        }
+
         Patient patient = Patient.findById(id);
+
+        if (patient == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity(ENTITY_NOT_FOUND).build();
+        }
+
         return Response.ok(patient).build();
     }
 
     @GET
     @Path("/find")
-    public List<Patient> findPatient(@QueryParam String name) {
-        return Patient.list("LOWER(name) LIKE ?1", "%" + name.toLowerCase() + "%");
+    public Response findPatient(@QueryParam String name) {
+        if (name.isBlank()) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
+
+        List<Patient> patientList = Patient.list("LOWER(name) LIKE ?1", "%" + name.toLowerCase() + "%");
+
+        if (patientList.isEmpty()) {
+            return Response.status(Response.Status.NOT_FOUND).entity(ENTITY_NOT_FOUND).build();
+        }
+
+        return Response.ok(patientList).build();
     }
 
     @GET
     @Path("/{id}/status")
-    public List<ReadingProjection> getPatientStatus(@Parameter(
-            description = "Number of records to be returned.",
-            required = true,
-            schema = @Schema(type = SchemaType.INTEGER)) @PathParam long id) {
+    public List<ReadingProjection> getPatientStatus(@PathParam long id) {
         List<Reading> patientStatus = patientService.getPatientStatus(id);
         return patientStatus.stream().map(ReadingProjection::new).collect(Collectors.toList());
     }
